Move odata-metadata-generator into odata-metadata directory
Change-type: patch
